ATR-821: renamed analyzer from PXAction on non primary view to the correct name - PXAction on non primary DAC
