+++ src/Acuminator/Acuminator.Tests/Acuminator.Tests.csproj
@@ -675,11 +675,11 @@
     <EmbeddedResource Include="Tests\StaticAnalysis\DacNonAbstractFieldType\Sources\SOOrderNotAbstractField.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\DacNonAbstractFieldType\Sources\SOOrderNotAbstractField_Expected.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\DacPropertyAttributes\Sources\DacFieldAttributesTypeMismatch_Expected.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryView\Sources\GraphWithNonPrimaryDacView.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryView\Sources\GraphWithNonPrimaryDacView_Expected.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryDac\Sources\GraphWithNonPrimaryDacView.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryDac\Sources\GraphWithNonPrimaryDacView_Expected.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\NoPrimaryViewForPrimaryDac\Sources\NoPrimaryViewForPrimaryDac.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryView\Sources\DerivedGraphWithBaseGraphPrimaryDac.cs" />
-    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryView\Sources\DerivedGraphWithBaseGraphPrimaryDac_Expected.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryDac\Sources\DerivedGraphWithBaseGraphPrimaryDac.cs" />
+    <EmbeddedResource Include="Tests\StaticAnalysis\PXActionOnNonPrimaryDac\Sources\DerivedGraphWithBaseGraphPrimaryDac_Expected.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\UnderscoresInDac\Sources\DacWithUnderscores.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\UnderscoresInDac\Sources\DacExtensionWithUnderscores.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\UnderscoresInDac\Sources\DacExtensionWithUnderscores_Expected.cs" />
@@ -797,7 +797,7 @@
     <EmbeddedResource Include="Tests\StaticAnalysis\InheritanceFromPXCacheExtension\Sources\InheritanceFromPXCacheExtension_Good.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\InheritanceFromPXCacheExtension\Sources\InheritanceFromPXCacheExtensionMakeSealed_Bad.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\InheritanceFromPXCacheExtension\Sources\InheritanceFromPXCacheExtensionMakeSealed_Bad_Expected.cs" />
-    <Compile Include="Tests\StaticAnalysis\PXActionOnNonPrimaryView\PXActionOnNonPrimaryViewTests.cs" />
+    <Compile Include="Tests\StaticAnalysis\PXActionOnNonPrimaryDac\PXActionOnNonPrimaryDacTests.cs" />
     <Compile Include="Tests\StaticAnalysis\StartRowResetForPaging\StartRowResetForPagingTests.cs" />
     <Compile Include="Tests\StaticAnalysis\ViewDeclarationOrder\ViewDeclarationOrderTests.cs" />
     <EmbeddedResource Include="Tests\StaticAnalysis\InvalidPXActionSignature\Sources\InvalidPXActionSignature.cs" />
