+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/PXActionOnNonPrimaryDac/PXActionOnNonPrimaryDacTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 
 using Acuminator.Analyzers.StaticAnalysis;
-using Acuminator.Analyzers.StaticAnalysis.PXActionOnNonPrimaryView;
+using Acuminator.Analyzers.StaticAnalysis.PXActionOnNonPrimaryDac;
 using Acuminator.Analyzers.StaticAnalysis.PXGraph;
 using Acuminator.Tests.Helpers;
 using Acuminator.Tests.Verification;
@@ -12,34 +12,34 @@ using Microsoft.CodeAnalysis.Diagnostics;
 
 using Xunit;
 
-namespace Acuminator.Tests.Tests.StaticAnalysis.PXActionOnNonPrimaryView
+namespace Acuminator.Tests.Tests.StaticAnalysis.PXActionOnNonPrimaryDac
 {
-	public class PXActionOnNonPrimaryViewTests : CodeFixVerifier
+	public class PXActionOnNonPrimaryDacTests : CodeFixVerifier
 	{
 		protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer() =>
 			new PXGraphAnalyzer(CodeAnalysisSettings.Default
 													.WithStaticAnalysisEnabled()
 													.WithSuppressionMechanismDisabled(),
-								new PXActionOnNonPrimaryViewAnalyzer());
+								new PXActionOnNonPrimaryDacAnalyzer());
 
-		protected override CodeFixProvider GetCSharpCodeFixProvider() => new PXActionOnNonPrimaryViewFix();
+		protected override CodeFixProvider GetCSharpCodeFixProvider() => new PXActionOnNonPrimaryDacFix();
 
 		[Theory]
 		[EmbeddedFileData("GraphWithNonPrimaryDacView.cs")] 
 		public virtual Task Test_Diagnostic_For_Graph_And_Graph_Extension(string source) =>
 			VerifyCSharpDiagnosticAsync(source,
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(23, 10, "Release1", "SOOrder"),
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(25, 10, "Release2", "SOOrder"),
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(34, 10, "Action1", "SOOrder"),
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(38, 10, "Action3", "SOOrder"),
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(47, 10, "Release1", "SOOrder"),
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(49, 10, "Release2", "SOOrder"));
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(23, 10, "Release1", "SOOrder"),
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(25, 10, "Release2", "SOOrder"),
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(34, 10, "Action1", "SOOrder"),
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(38, 10, "Action3", "SOOrder"),
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(47, 10, "Release1", "SOOrder"),
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(49, 10, "Release2", "SOOrder"));
 
 		[Theory]
 		[EmbeddedFileData("DerivedGraphWithBaseGraphPrimaryDac.cs")]
 		public virtual Task Test_Diagnostic_For_Derived_Graph(string source) =>
 			VerifyCSharpDiagnosticAsync(source,
-				Descriptors.PX1012_PXActionOnNonPrimaryView.CreateFor(26, 10, "Release1", "SOOrder"));
+				Descriptors.PX1012_PXActionOnNonPrimaryDac.CreateFor(26, 10, "Release1", "SOOrder"));
 
 		[Theory]
 		[EmbeddedFileData("GraphWithNonPrimaryDacView.cs",
